nlp-service/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic-settings>=2.2.0
spacy>=3.7.0
pymupdf>=1.24.0
python-docx>=1.1.0
python-multipart>=0.0.9
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
